feat: refactor streamlit_app.py to split streamlit and mcp part for modularization
